requirements.txt
numpy
pandas
matplotlib
scikit-image
pillow
tqdm
scikit-learn
psutil
datetime
scikit-image
albumentations